requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
redis
pytest
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
transformers
torch
locust